Create saltPepperNoise.py
Salt-and-pepper noise is a form of noise sometimes seen on images. It presents itself as sparsely occurring white and black pixels. An effective noise reduction method for this type of noise is a median filter or a morphological filter. This code adds salt and pepper noise where a pixel is turned into  dark (0) or white (255) with a probability of (p) for each to 'cap.bmp' image.